Rewrite schema and migrations for new context
As I am using now a context for the different parts I have rewriten
the schema and the migrations to include a prefix in the table
name. The prefix is an abbreviation of the context name.
One needs to run `mix ecto.drop` and `mix ecto.setup` after this change.